backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
minio>=7.2.0
redis>=5.0.0
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
httpx>=0.26.0
requests>=2.31.0
unstructured>=0.12.0
pypdf>=4.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.3
tiktoken>=0.5.0
openai>=1.10.0
Pillow>=10.0.0
pytesseract>=0.3.10
anthropic>=0.18.0
python-docx>=1.1.0
markdown>=3.5.0
weasyprint>=60.0
